+++ packages/integration-testsuite/tsconfig.json
@@ -8,5 +8,5 @@
     "outDir": "./dist"
   },
   "include": ["src/**/*"],
-  "references": []
+  "references": [{ "path": "../server" }]
 }
