+++ packages/server/src/__tests__/express/integration.test.ts
@@ -2,24 +2,15 @@ import { json } from 'body-parser';
 import cors from 'cors';
 import express from 'express';
 import http from 'http';
-import {
-  ApolloServer,
-  ApolloServerOptions,
-  BaseContext,
-  expressMiddleware,
-} from '../..';
+import { ApolloServer, expressMiddleware } from '../..';
 import { ApolloServerPluginDrainHttpServer } from '../../plugin/drainHttpServer';
 import { urlForHttpServer } from '../../utils/urlForHttpServer';
-import type {
-  CreateServerForIntegrationTestsOptions,
-  CreateServerForIntegrationTestsResult,
-} from '../integration';
 import { defineIntegrationTestSuite } from '@apollo/server-integration-testsuite';
 
 defineIntegrationTestSuite(async function (
-  serverOptions: ApolloServerOptions<BaseContext>,
-  testOptions?: CreateServerForIntegrationTestsOptions,
-): Promise<CreateServerForIntegrationTestsResult> {
+  serverOptions,
+  testOptions,
+) {
   const app = express();
   const httpServer = http.createServer(app);
   const server = new ApolloServer({
